+++ src/main/java/org/folio/inventory/domain/CollectionProvider.java
@@ -1,5 +1,6 @@
 package org.folio.inventory.domain;
 
+import org.folio.inventory.domain.authority.AuthorityCollection;
 import org.folio.inventory.domain.ingest.IngestJobCollection;
 import org.folio.inventory.domain.instances.InstanceCollection;
 import org.folio.inventory.domain.items.ItemCollection;
@@ -10,7 +11,7 @@ public interface CollectionProvider {
   HoldingCollection getHoldingCollection(String tenantId, String token);
   HoldingsRecordCollection getHoldingsRecordCollection(String tenantId, String token);
   InstanceCollection getInstanceCollection(String tenantId, String token);
+  AuthorityCollection getAuthorityCollection(String tenantId, String token);
   IngestJobCollection getIngestJobCollection(String tenantId, String token);
-  AuthorityRecordCollection getAuthorityRecordCollection(String tenantId, String token);
   UserCollection getUserCollection(String tenantId, String token);
 }