+++ src/main/java/org/folio/inventory/domain/authority/Authority.java
@@ -1,0 +1,4 @@
+package org.folio.inventory.domain.authority;
+
+public class Authority {
+}