+++ src/main/java/org/folio/inventory/domain/authority/AuthorityCollection.java
@@ -1,0 +1,8 @@
+package org.folio.inventory.domain.authority;
+
+import org.folio.Authority;
+import org.folio.inventory.domain.AsynchronousCollection;
+import org.folio.inventory.domain.SearchableCollection;
+
+public interface AuthorityCollection extends AsynchronousCollection<Authority>, SearchableCollection<Authority> {
+}